backend/api/requirements.txt
flask
Pillow
flask_cors
google-generativeai
python-dotenv
requests
